+++ app/src/main/java/edu/cmu/juicymeeting/database/model/Event.java
@@ -20,8 +20,6 @@ public class Event implements Parcelable {
     private String date;
     private long titleContextColor;
     private long imageContextColor;
-   // private boolean
-
 
 
     public String agenda;
