--- app/src/main/java/edu/cmu/juicymeeting/juicymeeting/OnItemClickListener.java
@@ -1,10 +1,0 @@
-package edu.cmu.juicymeeting.juicymeeting;
-
-import android.view.View;
-
-/**
- * Created by chenlinquan on 11/19/15.
- */
-public interface OnItemClickListener {
-    public void onItemClick(View view , int position);
-}
