+++ app/src/main/java/edu/cmu/juicymeeting/juicymeeting/activity/ChatroomActivity.java
@@ -1,4 +1,4 @@
-package edu.cmu.juicymeeting.juicymeeting;
+package edu.cmu.juicymeeting.juicymeeting.activity;
 
 import android.media.Ringtone;
 import android.media.RingtoneManager;
@@ -27,9 +27,10 @@ import java.util.Locale;
 import edu.cmu.juicymeeting.database.chatDB.DatabaseConnector;
 import edu.cmu.juicymeeting.database.chatDB.Group;
 import edu.cmu.juicymeeting.database.chatDB.Message;
+import edu.cmu.juicymeeting.juicymeeting.R;
 import edu.cmu.juicymeeting.util.ChatUtil;
-import edu.cmu.juicymeeting.util.MessageListAdapter;
-import edu.cmu.juicymeeting.util.WsConfig;
+import edu.cmu.juicymeeting.juicymeeting.adapter.MessageListAdapter;
+import edu.cmu.juicymeeting.ws.WsConfig;
 
 public class ChatroomActivity extends AppCompatActivity {
     public final static String TAG = ChatroomActivity.class.getSimpleName();
