--- app/src/main/java/edu/cmu/juicymeeting/util/ParserJson.java
@@ -1,92 +1,0 @@
-package edu.cmu.juicymeeting.util;
-
-import android.util.Log;
-
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-
-/**
- * Created by chenlinquan on 12/7/15.
- */
-public class ParserJson {
-    public static JSONArray getJSONFromURL(String url) {
-
-        // initialize
-        InputStream is = null;
-        String result = "";
-        JSONArray jObject = null;
-        InputStream inputStream = null;
-        // http post
-        try {
-            // create HttpClient
-            HttpClient httpclient = new DefaultHttpClient();
-
-            // make GET request to the given URL
-            HttpResponse httpResponse = httpclient.execute(new HttpGet(url));
-
-            // receive response as inputStream
-            inputStream = httpResponse.getEntity().getContent();
-
-            // convert inputstream to string
-            if(inputStream != null)
-                result = convertInputStreamToString(inputStream);
-            else
-                result = "Did not work!";
-
-//            HttpPost httppost = new HttpPost(url);
-//            HttpResponse response = httpclient.execute(httppost);
-//            HttpEntity entity = response.getEntity();
-//            is = entity.getContent();
-
-        } catch (Exception e) {
-            Log.e("log_tag", "Error in http connection " + e.toString());
-        }
-
-//        // convert response to string
-//        try {
-//            BufferedReader reader = new BufferedReader(new InputStreamReader(is, "iso-8859-1"), 8);
-//            StringBuilder sb = new StringBuilder();
-//            String line = null;
-//            while ((line = reader.readLine()) != null) {
-//                sb.append(line + "\n");
-//            }
-//            is.close();
-//            result = sb.toString();
-//        } catch (Exception e) {
-//            Log.e("log_tag", "Error converting result " + e.toString());
-//        }
-
-        // try parse the string to a JSON object
-        try {
-
-            jObject = new JSONArray(result);
-        } catch (JSONException e) {
-            Log.e("log_tag", "Error parsing data " + e.toString());
-        }
-        return jObject;
-    }
-
-    private static String convertInputStreamToString(InputStream inputStream) throws IOException {
-        BufferedReader bufferedReader = new BufferedReader( new InputStreamReader(inputStream));
-        String line = "";
-        String result = "";
-        while((line = bufferedReader.readLine()) != null)
-            result += line;
-
-        inputStream.close();
-        return result;
-
-    }
-}
