--- app/src/main/java/edu/cmu/juicymeeting/util/WsConfig.java
@@ -1,5 +1,0 @@
-package edu.cmu.juicymeeting.util;
-// 192.168.0.15
-public class WsConfig {
-	public static final String URL_WEBSOCKET = "ws://ec2-52-91-106-6.compute-1.amazonaws.com/WebMobileGroupChatServer/chat?";
-}
