+++ app/src/main/java/edu/cmu/juicymeeting/ws/WsConfig.java
@@ -1,0 +1,6 @@
+package edu.cmu.juicymeeting.ws;
+
+public class WsConfig {
+	public static final String URL_WEBSOCKET =
+			"ws://ec2-52-91-106-6.compute-1.amazonaws.com/WebMobileGroupChatServer/chat?";
+}
